finished updated manage accoutns stuff, updated tests too
