+++ tests/manage-accounts.test.ts
@@ -6,8 +6,7 @@ import { isValidSubstrateAddress } from '@entropyxyz/sdk/utils'
 import Keyring from '@entropyxyz/sdk/keys'
 import { randomAsHex } from '@polkadot/util-crypto'
 import { EntropyAccountConfig, EntropyConfig } from '../src/config/types'
-import { listAccounts } from '../src/flows/manage-accounts/list'
-import { createAccount } from '../src/flows/manage-accounts/helpers/create-account'
+import { createAccount, listAccounts } from '../src/accounts/utils'
 import * as config from '../src/config'
 import { promiseRunner, sleep } from './testing-utils'
 import { charlieStashAddress, charlieStashSeed } from './testing-utils/constants'
@@ -40,7 +39,7 @@ test('List Accounts', async t => {
   t.deepEqual(accountsArray, [{
     name: account.name,
     address: account.address,
-    verifyingKeys: account.data.admin.verifyingKeys
+    verifyingKeys: account?.data?.admin?.verifyingKeys
   }])
 
   // Resetting accounts on config to test for empty list
@@ -54,8 +53,6 @@ test('List Accounts', async t => {
 
   t.end()
 })
-
-const networkType = 'two-nodes'
 
 let counter = 0
 test('Create Account', async t => {
@@ -74,6 +71,6 @@ test('Create Account', async t => {
   const isValidAddress = isValidSubstrateAddress(newAccount.address)
 
   t.ok(isValidAddress, 'Valid address created')
-  t.equal(newAccount.address, admin.address, 'Generated Account matches Account created by Keyring')
+  t.equal(newAccount.address, admin?.address, 'Generated Account matches Account created by Keyring')
   t.end()
 })
